requirement.txt
# search
duckduckgo-search

# web service
beautifulsoup4
fastapi
uvicorn

# math
numpy

# ai
openai
langchain
torch
transformers
sentence-transformers
faiss-cpu
tiktoken


# output
colorama

# setings
python-dotenv

# api
pca
